Revert "Clean Fix (#16)" (#20)
This reverts commit f04ddd5d03e43db622cc9807ec2e27b572ea47f2.
