+++ src/Application/Services/DashboardService.js
@@ -1,10 +1,6 @@
-const BASE_URL = "https://localhost:5107/api/sensor";
-
-/**
- * Henter sensordata fra backend-API'et.
- */
+// Henter sensordata fra API'et
 export const fetchSensorData = async () => {
-  const response = await fetch(BASE_URL, {
+  const response = await fetch("https://localhost:5107/api/sensor", {
     credentials: "include",
   });
