--- src/Application/Services/SensorService.js
@@ -1,8 +1,0 @@
-import * as SensorAPI from "@/Infrastructure/API/SensorAPI";
-
-/**
- * Henter sensordata via SensorAPI – evt. klar til transformation eller caching.
- */
-export const fetchSensorData = async () => {
-  return await SensorAPI.fetchSensorData();
-};
